Make separate content API
